+++ library/core/src/main/java/com/google/android/exoplayer2/BasePlayer.java
@@ -94,11 +94,19 @@ public abstract class BasePlayer implements Player {
   @Override
   @Nullable
   public final Object getCurrentTag() {
-    int windowIndex = getCurrentWindowIndex();
     Timeline timeline = getCurrentTimeline();
-    return windowIndex >= timeline.getWindowCount()
+    return timeline.isEmpty()
         ? null
-        : timeline.getWindow(windowIndex, window, /* setTag= */ true).tag;
+        : timeline.getWindow(getCurrentWindowIndex(), window, /* setTag= */ true).tag;
+  }
+
+  @Override
+  @Nullable
+  public final Object getCurrentManifest() {
+    Timeline timeline = getCurrentTimeline();
+    return timeline.isEmpty()
+        ? null
+        : timeline.getWindow(getCurrentWindowIndex(), window, /* setTag= */ false).manifest;
   }
 
   @Override