+++ library/core/src/test/java/com/google/android/exoplayer2/source/ConcatenatingMediaSourceTest.java
@@ -226,7 +226,7 @@ public final class ConcatenatingMediaSourceTest {
     FakeMediaSource[] fastSources = createMediaSources(2);
     final FakeMediaSource[] lazySources = new FakeMediaSource[4];
     for (int i = 0; i < 4; i++) {
-      lazySources[i] = new FakeMediaSource(null, null);
+      lazySources[i] = new FakeMediaSource(null);
     }
 
     // Add lazy sources and normal sources before preparation. Also remove one lazy source again
@@ -307,16 +307,16 @@ public final class ConcatenatingMediaSourceTest {
     Timeline timeline = testRunner.prepareSource();
     TimelineAsserts.assertEmpty(timeline);
 
-    mediaSource.addMediaSource(new FakeMediaSource(Timeline.EMPTY, null));
+    mediaSource.addMediaSource(new FakeMediaSource(Timeline.EMPTY));
     timeline = testRunner.assertTimelineChangeBlocking();
     TimelineAsserts.assertEmpty(timeline);
 
     mediaSource.addMediaSources(
         Arrays.asList(
             new MediaSource[] {
-              new FakeMediaSource(Timeline.EMPTY, null), new FakeMediaSource(Timeline.EMPTY, null),
-              new FakeMediaSource(Timeline.EMPTY, null), new FakeMediaSource(Timeline.EMPTY, null),
-              new FakeMediaSource(Timeline.EMPTY, null), new FakeMediaSource(Timeline.EMPTY, null)
+              new FakeMediaSource(Timeline.EMPTY), new FakeMediaSource(Timeline.EMPTY),
+              new FakeMediaSource(Timeline.EMPTY), new FakeMediaSource(Timeline.EMPTY),
+              new FakeMediaSource(Timeline.EMPTY), new FakeMediaSource(Timeline.EMPTY)
             }));
     timeline = testRunner.assertTimelineChangeBlocking();
     TimelineAsserts.assertEmpty(timeline);
@@ -362,9 +362,9 @@ public final class ConcatenatingMediaSourceTest {
   public void testDynamicChangeOfEmptyTimelines() throws IOException {
     FakeMediaSource[] childSources =
         new FakeMediaSource[] {
-          new FakeMediaSource(Timeline.EMPTY, /* manifest= */ null),
-          new FakeMediaSource(Timeline.EMPTY, /* manifest= */ null),
-          new FakeMediaSource(Timeline.EMPTY, /* manifest= */ null),
+          new FakeMediaSource(Timeline.EMPTY),
+          new FakeMediaSource(Timeline.EMPTY),
+          new FakeMediaSource(Timeline.EMPTY),
         };
     Timeline nonEmptyTimeline = new FakeTimeline(/* windowCount = */ 1);
 
@@ -387,7 +387,7 @@ public final class ConcatenatingMediaSourceTest {
 
   @Test
   public void testIllegalArguments() {
-    MediaSource validSource = new FakeMediaSource(createFakeTimeline(1), null);
+    MediaSource validSource = new FakeMediaSource(createFakeTimeline(1));
 
     // Null sources.
     try {
@@ -660,8 +660,8 @@ public final class ConcatenatingMediaSourceTest {
                 10 * C.MICROS_PER_SECOND,
                 FakeTimeline.createAdPlaybackState(
                     /* adsPerAdGroup= */ 1, /* adGroupTimesUs= */ 0)));
-    FakeMediaSource mediaSourceContentOnly = new FakeMediaSource(timelineContentOnly, null);
-    FakeMediaSource mediaSourceWithAds = new FakeMediaSource(timelineWithAds, null);
+    FakeMediaSource mediaSourceContentOnly = new FakeMediaSource(timelineContentOnly);
+    FakeMediaSource mediaSourceWithAds = new FakeMediaSource(timelineWithAds);
     mediaSource.addMediaSource(mediaSourceContentOnly);
     mediaSource.addMediaSource(mediaSourceWithAds);
 
@@ -807,7 +807,7 @@ public final class ConcatenatingMediaSourceTest {
   @Test
   public void testDuplicateMediaSources() throws IOException, InterruptedException {
     Timeline childTimeline = new FakeTimeline(/* windowCount= */ 2);
-    FakeMediaSource childSource = new FakeMediaSource(childTimeline, /* manifest= */ null);
+    FakeMediaSource childSource = new FakeMediaSource(childTimeline);
 
     mediaSource.addMediaSource(childSource);
     mediaSource.addMediaSource(childSource);
@@ -840,7 +840,7 @@ public final class ConcatenatingMediaSourceTest {
   @Test
   public void testDuplicateNestedMediaSources() throws IOException, InterruptedException {
     Timeline childTimeline = new FakeTimeline(/* windowCount= */ 1);
-    FakeMediaSource childSource = new FakeMediaSource(childTimeline, /* manifest= */ null);
+    FakeMediaSource childSource = new FakeMediaSource(childTimeline);
     ConcatenatingMediaSource nestedConcatenation = new ConcatenatingMediaSource();
 
     testRunner.prepareSource();
@@ -874,8 +874,7 @@ public final class ConcatenatingMediaSourceTest {
   public void testClear() throws IOException {
     DummyMainThread dummyMainThread = new DummyMainThread();
     final FakeMediaSource preparedChildSource = createFakeMediaSource();
-    final FakeMediaSource unpreparedChildSource =
-        new FakeMediaSource(/* timeline= */ null, /* manifest= */ null);
+    final FakeMediaSource unpreparedChildSource = new FakeMediaSource(/* timeline= */ null);
     dummyMainThread.runOnMainThread(
         () -> {
           mediaSource.addMediaSource(preparedChildSource);
@@ -1092,13 +1091,13 @@ public final class ConcatenatingMediaSourceTest {
   private static FakeMediaSource[] createMediaSources(int count) {
     FakeMediaSource[] sources = new FakeMediaSource[count];
     for (int i = 0; i < count; i++) {
-      sources[i] = new FakeMediaSource(createFakeTimeline(i), null);
+      sources[i] = new FakeMediaSource(createFakeTimeline(i));
     }
     return sources;
   }
 
   private static FakeMediaSource createFakeMediaSource() {
-    return new FakeMediaSource(createFakeTimeline(/* index */ 0), null);
+    return new FakeMediaSource(createFakeTimeline(/* index */ 0));
   }
 
   private static FakeTimeline createFakeTimeline(int index) {