+++ library/ui/src/main/java/com/google/android/exoplayer2/ui/PlayerControlView.java
@@ -1212,8 +1212,7 @@ public class PlayerControlView extends FrameLayout {
     }
 
     @Override
-    public void onTimelineChanged(
-        Timeline timeline, @Nullable Object manifest, @Player.TimelineChangeReason int reason) {
+    public void onTimelineChanged(Timeline timeline, @Player.TimelineChangeReason int reason) {
       updateNavigation();
       updateTimeline();
     }