+++ testutils_robolectric/src/main/java/com/google/android/exoplayer2/testutil/StubExoPlayer.java
@@ -196,11 +196,6 @@ public abstract class StubExoPlayer extends BasePlayer implements ExoPlayer {
   }
 
   @Override
-  public Object getCurrentManifest() {
-    throw new UnsupportedOperationException();
-  }
-
-  @Override
   public Timeline getCurrentTimeline() {
     throw new UnsupportedOperationException();
   }