+++ RELEASENOTES.md
@@ -15,6 +15,9 @@
 * Add VR player demo.
 * Wrap decoder exceptions in a new `DecoderException` class and report as
   renderer error.
+* Do not pass the manifest to callbacks of Player.EventListener and
+  SourceInfoRefreshListener anymore. Instead make it accessible through
+  Player.getCurrentManifest() and Timeline.Window.manifest.
 
 ### 2.10.3 ###